+++ ccd/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.20.3"
+__version__ = "0.20.4"
 
 __all__ = ["__version__"]